pyproject.toml
[project]
name = "pulseengine"
version = "0.1.0"
description = "PulseEngine — real-time content ranking backend"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2,<7.0",
    "djangorestframework>=3.15,<4.0",
    "psycopg[binary]>=3.1,<4.0",
    "redis>=5.0,<6.0",
    "celery[redis]>=5.4,<6.0",
    "elasticsearch-dsl>=8.13,<9.0",
    "python-dotenv>=1.2,<2.0",
    "gunicorn>=22.0,<23.0",
    "dj-database-url>=2.1,<3.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.15.4",
]


[tool.ruff]
line-length = 100
